+++ app/controllers/solid_queue_web/processes_controller.rb
@@ -1,0 +1,7 @@
+module SolidQueueWeb
+  class ProcessesController < ApplicationController
+    def index
+      @processes = SolidQueue::Process.order(:kind, :name).to_a
+    end
+  end
+end